Create 'optionalDataMRZ()' as a module import.
