+++ modules/icao9303/MRVADocument.js
@@ -7,6 +7,7 @@ import { DEFAULT_PHOTO, DEFAULT_SIGNATURE_IMAGE } from "./utilities/default-imag
 import { getFullYearFromString } from "./utilities/get-full-year-from-string.js";
 import { generateMRZCheckDigit } from "./utilities/generate-mrz-check-digit.js";
 import { fullNameMRZ } from "./utilities/full-name-mrz.js";
+import { optionalDataMRZ } from "./utilities/optional-data-mrz.js";
 
 /**
  * Stores properties and methods for machine-readable visas (MRV-A) with
@@ -420,7 +421,7 @@ class MRVADocument {
     return fullNameMRZ(this, 39);
   }
   static #optionalDataToMRZ = function() {
-    return TravelDocument.optionalDataMRZ(this, 16);
+    return optionalDataMRZ(this, 16);
   }
 }
