+++ modules/icao9303/TD1Document.js
@@ -6,6 +6,7 @@ import { DEFAULT_PHOTO, DEFAULT_SIGNATURE_IMAGE } from "./utilities/default-imag
 import { getFullYearFromString } from "./utilities/get-full-year-from-string.js";
 import { generateMRZCheckDigit } from "./utilities/generate-mrz-check-digit.js";
 import { fullNameMRZ } from "./utilities/full-name-mrz.js";
+import { optionalDataMRZ } from "./utilities/optional-data-mrz.js";
 
 /**
  * Stores properties and methods for TD1-sized machine-readable travel documents
@@ -364,7 +365,7 @@ class TD1Document {
     return fullNameMRZ(this, 30);
   }
   static #optionalDataToMRZ = function() {
-    return TravelDocument.optionalDataMRZ(this, 26);
+    return optionalDataMRZ(this, 26);
   }
 }
