+++ modules/icao9303/TD2Document.js
@@ -6,6 +6,7 @@ import { DEFAULT_PHOTO, DEFAULT_SIGNATURE_IMAGE } from "./utilities/default-imag
 import { getFullYearFromString } from "./utilities/get-full-year-from-string.js";
 import { generateMRZCheckDigit } from "./utilities/generate-mrz-check-digit.js";
 import { fullNameMRZ } from "./utilities/full-name-mrz.js";
+import { optionalDataMRZ } from "./utilities/optional-data-mrz.js";
 
 /**
  * Stores properties and methods for TD2-sized machine-readable travel documents
@@ -340,7 +341,7 @@ class TD2Document {
     return fullNameMRZ(this, 31);
   }
   static #optionalDataToMRZ = function() {
-    return TravelDocument.optionalDataMRZ(this, 7);
+    return optionalDataMRZ(this, 7);
   }
 }
