+++ modules/icao9303/TD3Document.js
@@ -6,6 +6,7 @@ import { DEFAULT_PHOTO, DEFAULT_SIGNATURE_IMAGE } from "./utilities/default-imag
 import { getFullYearFromString } from "./utilities/get-full-year-from-string.js";
 import { generateMRZCheckDigit } from "./utilities/generate-mrz-check-digit.js";
 import { fullNameMRZ } from "./utilities/full-name-mrz.js";
+import { optionalDataMRZ } from "./utilities/optional-data-mrz.js";
 
 /**
  * Stores properties and methods for TD3-sized machine-readable travel documents
@@ -352,7 +353,7 @@ class TD3Document {
     return fullNameMRZ(this, 39);
   }
   static #optionalDataToMRZ = function() {
-    return TravelDocument.optionalDataMRZ(this, 14);
+    return optionalDataMRZ(this, 14);
   }
 }
